refactor(solitaire): isolate stats persistence
